fix bug about integer overflow

+++ inc/ConstraintTable.h
@@ -10,20 +10,20 @@ public:
 	int length_max = INT_MAX;
 	int goal_location;
 	int latest_timestep = 0; // No negative constraints after this timestep.
-	int num_col;
-	int map_size;
+	size_t num_col;
+	size_t map_size;
 
 	int getHoldingTime(); // the earliest timestep that the agent can hold its goal location
 
 	// void clear(){ct.clear(); cat_small.clear(); cat_large.clear(); landmarks.clear(); length_min = 0, length_max = INT_MAX; latest_timestep = 0;}
 
-	bool constrained(int loc, int t) const;
-    bool constrained(int curr_loc, int next_loc, int next_t) const;
-	int getNumOfConflictsForStep(int curr_id, int next_id, int next_timestep) const;
+	bool constrained(size_t loc, int t) const;
+    bool constrained(size_t curr_loc, size_t next_loc, int next_t) const;
+	int getNumOfConflictsForStep(size_t curr_id, size_t next_id, int next_timestep) const;
 
 
 	ConstraintTable() = default;
-	ConstraintTable(int num_col, int map_size, int goal_location = -1) : goal_location(goal_location), num_col(num_col), map_size(map_size) {}
+	ConstraintTable(size_t num_col, size_t map_size, int goal_location = -1) : goal_location(goal_location), num_col(num_col), map_size(map_size) {}
 	ConstraintTable(const ConstraintTable& other) {copy(other); }
 
 	void copy(const ConstraintTable& other);
@@ -35,15 +35,15 @@ protected:
 
 	unordered_map<size_t, size_t> landmarks; // <timestep, location>: the agent must be at the given location at the given timestep
 
-	void insert2CT(int loc, int t_min, int t_max); // insert a vertex constraint to the constraint table
-	void insert2CT(int from, int to, int t_min, int t_max); // insert an edge constraint to the constraint table
-	void insertLandmark(int loc, int t); // insert a landmark, i.e., the agent has to be at the given location at the given timestep
+	void insert2CT(size_t loc, int t_min, int t_max); // insert a vertex constraint to the constraint table
+	void insert2CT(size_t from, size_t to, int t_min, int t_max); // insert an edge constraint to the constraint table
+	void insertLandmark(size_t loc, int t); // insert a landmark, i.e., the agent has to be at the given location at the given timestep
 
 
-	inline int getEdgeIndex(int from,int to) const {return (1 + from) * map_size + to;  }
+	inline size_t getEdgeIndex(size_t from, size_t to) const { return (1 + from) * map_size + to; }
 
 private:
-	int map_size_threshold = 10000;
+	size_t map_size_threshold = 10000;
 	vector<list<size_t> > cat_large; // conflict avoidance table for large maps
 	vector<vector<size_t> > cat_small; // conflict avoidance table for small maps
