+++ inc/ReservationTable.h
@@ -11,16 +11,16 @@ public:
     double runtime;
 
 	ReservationTable() = default;
-	ReservationTable(int num_col, int map_size, int goal_location = -1): ConstraintTable(num_col, map_size, goal_location) {}
+	ReservationTable(size_t num_col, size_t map_size, int goal_location = -1): ConstraintTable(num_col, map_size, goal_location) {}
 	ReservationTable(const ConstraintTable& other) {copy(other); }
 
 
-    list<Interval> get_safe_intervals(int location, size_t lower_bound, size_t upper_bound);
-	list<Interval> get_safe_intervals(int from, int to, size_t lower_bound, size_t upper_bound);
+    list<Interval> get_safe_intervals(size_t location, size_t lower_bound, size_t upper_bound);
+	list<Interval> get_safe_intervals(size_t from, size_t to, size_t lower_bound, size_t upper_bound);
 
 	// int get_holding_time(int location);
-   Interval get_first_safe_interval(int location);
-    bool find_safe_interval(Interval& interval, int location, int t_min);
+   Interval get_first_safe_interval(size_t location);
+    bool find_safe_interval(Interval& interval, size_t location, size_t t_min);
 
 	void buildCAT(int agent, const vector<Path*>& paths); // build the conflict avoidance table
 
@@ -29,13 +29,13 @@ public:
 private:
 	unordered_map<size_t, list<Interval > > rt; // location -> [t_min, t_max), num_of_collisions
 
-	unordered_map< size_t, list<pair<int, int> > > cat; // conflict avoidance table cat:  location -> time range, or edge -> time range
+	unordered_map< size_t, list<pair<size_t, int> > > cat; // conflict avoidance table cat:  location -> time range, or edge -> time range
 
-    void insert2RT(int location, size_t t_min, size_t t_max);
-    void insertSoftConstraint2RT(int location, size_t t_min, size_t t_max);
+    void insert2RT(size_t location, size_t t_min, size_t t_max);
+    void insertSoftConstraint2RT(size_t location, size_t t_min, size_t t_max);
 	void mergeIntervals(list<Interval >& intervals);
 
-	void updateRT(int location); // update RT at the gvien location
+	void updateRT(size_t location); // update RT at the gvien location
 
-	int getNumOfConflictsForStep(int curr_id, int next_id, int next_timestep) const;
+	int getNumOfConflictsForStep(size_t curr_id, size_t next_id, size_t next_timestep) const;
 };
